move highlighting and clearing highlighting to its own js file
